Moved functionality to circuits folder for import dependencies
